requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
neo4j>=5.17.0
python-dotenv>=1.0.0
